Separate Firebase code and move it to impl package
