+++ app/src/main/AndroidManifest.xml
@@ -151,7 +151,7 @@
                 android:name="com.yalantis.ucrop.UCropActivity"
                 android:screenOrientation="portrait"
                 android:theme="@style/Theme.AppCompat.Light.NoActionBar"/>
-        <service android:name=".notifications.FirebaseNotificationService" >
+        <service android:name=".repository.impl.FirebaseNotificationService" >
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
             </intent-filter>
